+++ example/lib/startback_todo/operation.rb
@@ -3,3 +3,4 @@ module StartbackTodo
   end
 end
 require_relative "operation/create_todo"
+require_relative "operation/rate_limited"