+++ example/lib/startback_todo/operation/rate_limited.rb
@@ -1,0 +1,15 @@
+module StartbackTodo
+  class RateLimited < Operation
+
+    rate_limit
+
+    def initialize(input)
+      @input = input
+    end
+
+    def call
+      true
+    end
+
+  end
+end