+++ lib/startback/security/ext/operation.rb
@@ -1,0 +1,5 @@
+module Startback
+  class Operation
+    extend Security::RateLimit
+  end # class Operation
+end # module Startback