+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * Added SpyLogger that helps integration tests wanting to lightly observe
   call flows.
 
+* Added Startback::Security::RateLimiter to help limiting operation executions
+  as a security measure.
+
 ## 1.1.0 - 2025-02-09
 
 * Added ruby 3.3 to the build & release matrix